Updated tag format test location
